Forgot to change convention for modelRepresentation as well
If we consider the z-axis as being vertical and to be consistent with
the new convention of the camera matrices, we have to consider the

1. column in the obj. file as Y-axis
2. column in the obj. file as Z-axis
3. column in the obj. file as X-axis

I guess the same holds not only for the coordinates of the vertices but
also for the normals. Haven't tested this yet, but it seems very
unlikely that they use different conventions in the same .obj file.
Still, we need to be careful